Move properties
